+++ apps/bart/tailwind.config.ts
@@ -3,14 +3,14 @@ import {Config} from 'tailwindcss'
 import config from '@rubriclab/tailwind-config'
 
 const tailwindConfig = {
-	content: ['./app/**/*.tsx'],
+	content: ['./app/*.tsx'],
 	presets: [config],
 	theme: {
+		colors: {
+			...config.theme.colors,
+			'simpson-yellow': '#FDD41E'
+		},
 		extend: {
-			colors: {
-				...config.theme.extend.colors,
-				'simpson-yellow': '#FDD41E'
-			},
 			fontFamily: {
 				simpson: ['var(--font-simpson)']
 			}