+++ apps/web/app/blog/page.tsx
@@ -1,4 +1,5 @@
-import {ComingSoonButton, Header} from '@rubriclab/ui'
+import {Header} from '@rubriclab/ui'
+import {ComingSoonButton} from '@rubriclab/ui/src'
 import {ArrowUpRightIcon} from 'lucide-react'
 import {Metadata} from 'next'
 import Image from 'next/image'