+++ packages/@rubriclab/ui/src/ComingSoonButton.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import {toast} from './ToastProvider'
+import './globals.css'
 
 export const ComingSoonButton = () => {
 	return (