+++ packages/@rubriclab/ui/src/ContactButton.tsx
@@ -2,6 +2,7 @@
 
 import {copyToClipboard} from '@rubriclab/utils'
 import {toast} from './ToastProvider'
+import './globals.css'
 
 export const ContactButton = ({body}: {body: string}) => {
 	return (