+++ packages/@rubriclab/ui/src/Header.tsx
@@ -1,3 +1,7 @@
+'use client'
+
+import './globals.css'
+
 export const Header = ({text}: {text: string}) => {
 	return <h1>{text}</h1>
 }