+++ packages/@rubriclab/ui/tailwind.config.ts
@@ -3,15 +3,8 @@ import {Config} from 'tailwindcss'
 import config from '@rubriclab/tailwind-config'
 
 const tailwindConfig = {
-	content: ['./*.tsx'],
-	presets: [config],
-	theme: {
-		extend: {
-			colors: {
-				...config.theme.extend.colors
-			}
-		}
-	}
+	content: ['./src/*.tsx'],
+	presets: [config]
 } satisfies Config
 
 export default tailwindConfig