+++ packages/create-rubric-app/templates/fullstack/app/page.tsx
@@ -1,4 +1,5 @@
-import {ContactButton, Header} from '@rubriclab/ui'
+import {Header} from '@rubriclab/ui'
+import {ContactButton} from '@rubriclab/ui/src'
 import {Metadata} from 'next'
 import {DEFAULT_META} from '../constants/metadata'