+++ packages/@rubriclab/tailwind-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rubriclab/tailwind-config
 
+## 0.2.5
+
+### Patch Changes
+
+- Better tailwind configs
+
+## 0.2.4
+
+### Patch Changes
+
+- more tailwind fixes
+
+## 0.2.3
+
+### Patch Changes
+
+- more fixes to tailwind
+
 ## 0.2.2
 
 ### Patch Changes